pyproject.toml
[project]
name = "backend-dl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "tensorflow>=2.21.0",
    "uvicorn>=0.41.0",
]
